feat: transliterate user defined content (#359)
Use `jsii-rosetta` to transliterate assemblies before constructing the reflect TypeSystem.

This makes it so README code snippets as well as `@example` code blocks for inline documentation are also transliterated, providing a full language specific page. 

> Currently still in draft because it uses an unpublished version of `@jsii/spec` and `jsii-rosetta` pending a jsii release.

### Implementation Notes

Previously, the `Documentation` class accepted a **required** `reflect.Assembly` property. This meant that the creation of the type-system was the responsibility of the caller. This API is somewhat cumbersome,  and now that the assemblies need to be transliterated beforehand, this is even more so. 

The PR changes the input to the `Documentation` class, so that it only requires local paths that contain assembly files, it then proceeds to transliterate those assemblies and create the type-system on its own. 

Since the creation of the type-system is an `async` process, the `Documentation` class offers a `static async` function that needs to be awaited on, instead of directly using the constructor, which is now private. 

So the usage now becomes:

```ts
const docs = await Documentation.forLocalPackage(rootPath, assembliesPath);
const page = docs.render().render();
```

In addition, a new functionality of generating documentation for remote packages was added:

```ts
const docs = await Documentation.forRemotePackage('@aws-cdk/aws-ecr', '1.110.1');
const page = docs.render().render();
```

BREAKING CHANGE: The `Documentation` class can no longer be directly instantiated, use static factory methods instead.