feat: json schema rendering
Defines a schema in `docgen/schema` that applications can use to render
the api reference of a transliterated package. Adds a `renderToJson`
method on the `Documentation` class and relevant types.

This allows the construct-hub-webapp to further control how the api
reference is rendered and styled without needing to regenerate markdown
documents.