fix: docgen fails when installing from a custom registry defined in the HOME directory (#368)
The current implementation overrides the `HOME` env variable before running any `npm` command because otherwise running it inside a lambda function (where `HOME` is a read-only filesystem) will fail. 

This works well when installing from the default registry, but when a custom registry is configured in the actual `HOME` directory (via `.npmrc` file), this causes the command to ignore those configurations. 

As a result, and following https://github.com/cdklabs/construct-hub/pull/164, docgen generation currently fails because it tries to connect to the default registry, which it cannot since the lambda doesn't have internet access.

The solution is to completely drop the logic for overriding the `HOME` directory - it is not the responsibility of `jsii-docgen`, but rather of the consumer. 

In addition, this PR includes:

- Expose some needed information on `TransliteratedType` so that consumers can construct custom links using [`linkFormatter`](https://github.com/cdklabs/jsii-docgen/blob/master/src/docgen/view/documentation.ts#L47)
- Fix possible FQN collision between a type and a property/parameter. For example: the `@aws-cdk/aws-esk.AwsAuthProps` struct will have the same FQN as the `@aws-cdk/aws-eks.AwsAuth.props` parameter (after sanitation). 

For some context, all of these fixes are necessary for https://github.com/cdklabs/construct-hub/pull/175