feat: improve signalling of ENOSPC during "npm install"
Allow consumer code to react differently to failures caused by running out of
disk space from other classes of failures. For example, this can be used to cut
retries when the disk is full.

This is done by recognizing the exit code returned by `npm install` when it hits
`ENOSPC` and throwing a dedicated exception type in this scenario, so that
consumers can catch, then `instanceof` test it.