chore: use `npm --json` mode (#476)
In order to improve the reliability of error matching, use `npm` with
the `--json` flag, which causes a JSON object to be the only thing
output to `STDOUT`, which describes the result of the task (including
error details if it failed). This is a lot more elegant than the log
parsing we were previously doing.

Additionally, check the version of the `npm` that is on `$PATH` using
`npm version --json`, and use it if it is `npm >= v7`. This removes the
need to install `npm@8` in scenarios where the platform provides
sufficient tooling, which results in an increase of performance and
reliability.