feat: reuse package installation for multiple languages rendering (#496)
Currently the documentation API allows creating a `Documentation` object only for a specific language. This means that if callers are interested in rendering multiple languages, they have to either:

1. Call `Documentation.forPackage` multiple times, resulting in un-necessary package installations.
2. Install the package externally, and use `Documentation.forProject`. This means the package installation code has to be repeated by the caller. This is [exactly](https://github.com/cdklabs/construct-hub/blob/b289fd81ede5d7c5795a76d067ba38ebd8ba1aec/src/backend/transliterator/transliterator.ecstask.ts#L95-L105) what happens in `construct-hub`, which is a shame.

This PR allows reusing package installation for multiple languages by moving the `language` property to the `render` function, and preserving the package installation across calls. 

BREAKING CHANGE: The `language` and `loose` properties have moved from the `Documentation.forXXX` functions to the `Documentation.render` method.