fix: invalid values for submodule in JSON output (#571)
Fixes #570

The root cause of the bug was that when calling `toJson` in `transpile.ts`, we were generating the value of `submodule` from `TranspiledType`, which actually refers to the _transliterated_ submodule name. We fix this by creating a new field called `submoduleFqn` that stores the correct submodule name, and using that instead.

I've added extra unit tests with additional package + language combinations that previously failed to prevent regressions.

This PR also fixes a bug where transliterating a corrupted assembly did not properly throw `CorruptedAssemblyError`. The root cause of this was that we were doing some transpiling within the `toJson()` methods instead of within class constructors, and our try-catch block only covered the initialization of the `ApiReference` class. I fixed this by refactoring so all transpiling happens within constructors. (This was already the case for most classes, there were just a few exceptions where this wasn't being followed).