fix: language support validation is too shallow (#636)
The language support validation only checked for the presence of a
particular key under `jsii.targets`, however it failed to validate that
the contents of that key is valid, and did not throw a
`LanguageNotSupportedError`, instead letting `jsii-rosetta` throw a
generic error once it tried to use the missing configuration.

This adds in-depth checkers that, although they don't ensure the values
are valid, at least verifies they are present and of the expected type.