fix: header limit reached on some packages with examples (#641)
On some packages, we are seeing an error where jsii-docgen throws an error

```Error: Unable to render markdown. Header limit (6) reached.
    at MarkdownDocument.render (/Volumes/workplace/jsii-docgen/lib/docgen/render/markdown-doc.js:105:19)
```

This is caused by the existence of examples on parameters of methods in some libraries. To see concretely the chain of header sizes, I found that for `@aws-cdk/aws-apigateway@1.47.0` it was trying to render:

```
# API Reference <a name="API Reference" id="api-reference"></a>

## Constructs <a name="Constructs" id="Constructs"></a>

### ProxyResource <a name="ProxyResource" id="@aws-cdk/aws-apigateway.ProxyResource"></a>

#### Methods <a name="Methods" id="Methods"></a>

##### `add_method` <a name="add_method" id="@aws-cdk/aws-apigateway.ProxyResource.addMethod"></a>

###### `request_models`<sup>Optional</sup> <a name="request_models" id="@aws-cdk/aws-apigateway.ProxyResource.addMethod.parameter.requestModels"></a>

####### Example <a name="Example" id="@aws-cdk/aws-apigateway.ProxyResource.addMethod.parameter.requestModels.example"></a>
```

Since H7 is not allowed, we change examples to be rendered without a header, and instead with the title "Example" italicized.