Go formatting and enum member name fix
Removes unecessary `;` in go callable formatting and adds return types
where needed on static functions and on initializers to be closer to
idiomatic Go docs. Changes struct member signatures to not be function
calls when they are properties.

Fixes bug in JSON generation of the enum member types which displayed
non-transliterated names.