fix: transliterate snippets missing from tablet (#651)
Catch failures of `transliterateAssembly` and rethrow as
`LanguageNotSupportedError`. This allows adding support for Go as some
package versions were released with Go support before jsii-rosetta
supported it and have a tablet file without Go snippets.

In the future live transliteration of these missing snippets may be
ideal but the only package I've found with this scenario is aws-cdk-lib,
which takes a really long time to live transliterate and is something we
may want to optimize further before enabling.

Furthermore only a handful of aws-cdk-lib versions actually have this
problem as the latest have correct Go snippets in the tablet, andd the
earliest dont' have tablet files at all and are therefore fully transliterated
by jsii-docgen.