fix: detect packages removed from npm as "uninstallable"
If a package is deleted from npm and you try to install it you will
receive this error:

```
{
  "error": {
    "code": null,
    "summary": "Cannot convert undefined or null to object",
    "detail": ""
  }
}
```

Detect that error as an "uninstallable" package. This is a pretty
non-specific error message so the only risk is that this error message
overlaps with a different failure. Hopefully an actual error would have
the code and detail populated.