fix: return a `LanguageNotSupportedError` instead of a generic error (#914)
Construct Hub uses the type of error to determine what action to take. If the `transliterator` task fails to process a particular language it checks to see if the error is a `LanguageNotSupportedError` and if so it handles the error by marking the language as not supported instead of throwing the error. See

https://github.com/cdklabs/construct-hub/blob/55e75fa4bf289832d1fe0d2717bd670b5f650bb3/src/backend/transliterator/transliterator.ecstask.ts#L267

Fixes #