fix: support hoisted dependencies (#981)
This PR is a updated revision of: https://github.com/cdklabs/jsii-docgen/pull/644

Removes copy of the project directory and instead only outputs
transliterated assemblies to the working directory. This avoids
potential problems when generating docs for packages that may have some
dependencies that are symlinks, such as those in a monorepo.

Fixes #390