fix: update glob-promise to fix an issue with an incorrect peer dependency
In glob@9, the glob package removed the callback option as it now supports
promises instead. However, glob-promise had a peer dependency of '*', which meant
it was possible for it to use 9.x and later - resulting in no output from
jsii-docgen, as the callback was never called.

This commit updates glob-promise to a version with an appropriate peer
dependency, and updates the version of glob to one which matches the peer dependency.