fix: copy failures when git's core.fsmonitor is enabled (#1045)
The doc-gen process copies the current project directory to a temporary directory, including a potential `.git` directory. If `code.fsmonitor` is enabled, this directory may contain a socket file, which cannot be copied.

This adds a filter function to the `fs-extra.copy` call that excludes any non-regular file (i.e: file or directory), so that it no longer fails in that situation.