feat!: upgrade jsii-rosetta to 5.1.x and make it a peer dependency (#1042)
BREAKING CHANGE: Change `jsii-rosetta` to be a peer dependency with a tested version range. Minimum version is `5.1.2` with this release and we will add new minor versions once they are tested.

Package managers will attempt to install the peer dependency, unless explicitly configured not to.
You can install newer versions of `jsii-rosetta` when they are released, but they should be considered untested until the peer dependency version range is updated.