fix: unrelated corrupted assembly causes failure (#1155)
Docgen will traverse the directory tree to speculative load any found jsii assemblies into the TypeSystem.
While doing this, it checks if the found assembly is the assembly that it actually wants to generate docs for.
Only the name of the found assembly is needed for this check, since the actual loading happens later on.
 
This check incorrectly validated all assembly files, causing the whole generation to fail when an invalid or otherwise corrupted assembly was encounterd, even when this assembly was completely unrelated.