fix: split by submodule ignores output (#1159)
Previously submodule files used to be created in the project root, no matter what value `--output` had.
This also caused invalid links in the index files, which was correctly placed at the specified output location, but links assumed the submodule files to be in the same directory.

Also fixes the command failing when the output directory does not exists.