feat: add support for class-covariant-overrides jsii feature (#1849)
This PR adds support for the `class-covariant-overrides` jsii feature to
the list of supported assembly features.

This change doesn't require any code modifications beyond adding the
feature name to the supported list. The existing jsii-docgen
infrastructure already handles this feature type correctly - we just
need to explicitly declare support for it to prevent compatibility
warnings.

Fixes #